Split `PNExpressionArithmetic` into separate file
