+++ src/core/core.jl
@@ -3,5 +3,6 @@ include("utilities/truncated_series_monoid.jl")
 include("utilities/truncated_series_inversion.jl")
 include("PNTerm.jl")
 include("PNExpansion.jl")
+include("PNExpressionArithmetic.jl")
 include("PNExpression.jl")
 include("PNReference.jl")
